finished mfing user prof event index display
