+++ frontend/src/util/event_api_util.js
@@ -20,12 +20,7 @@ export const deleteEvent = (eventId) => {
   return axios.delete(`api/events/${eventId}`);
 };
 
-export const getUserEvents = (id) => {
-  return axios.get(`/api/user/${id}`);
-};
-
 export const joinEvent = (eventId, user) => {
-  // debugger;
   return axios.patch(
     // `http://localhost:5000/api/events/${eventId}/add_attendee`,
     `api/events/${eventId}/add_attendee`,
